requirements.txt
flask
pdfminer.six
openai